Saving and updating with sections works
Was digging myself into a hole and restarted the FE-BE integration process, trying to only use graphql interactions when saving a lesson.
Used ObjectIDs in FE as converting uuids from FE to ObjectIDs in the BE was not going well.
Updated LessonCreateView to use GET_LESSON instead of fetchLesson
Update Lesson schema to have sections as an array of Sections whereas prior it was an array of SectionIDs
Updated LessonCreateView, FileMenuManager and useSaveLesson to save and update lessons with sections (need to work on other fields still)